add bf16 infer
